Rename files to use UpperCamelCaseCluster/Logic for soil-measurement-server
